Directly fix boost's shared_ptr.hpp for deleted constructor
Solution on Stackoverflow https://stackoverflow.com/a/18900875/571227
which suggests to update boost header to version 1.48+ or directly
modify to add such constructor.

This changes opt-in to directly add the changes to not risk making
semantics changes to other sub-projects although boost::shared_ptr
is only used in copy_on_write sub-project.